feat: SettlementDialogContent 컴포넌트에 onSubmit prop 추가 및 적용
